pyproject.toml
[tool.poetry]
name = "onlineshop"
version = "0.1.0"
description = ""
authors = ["Lopkop <lopkop@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-environ = "^0.4.5"
django-crispy-forms = "^1.12.0"
cloudipsp = "^1.0.2"
django-recaptcha = "^2.0.6"
psycopg2 = "^2.9.1"
selenium = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
